+++ src/game-coordinator/services/game-runtime.service.ts
@@ -36,6 +36,9 @@ export class GameRuntimeService implements OnModuleInit {
       async ({ gameId, replaceeId, replacementId }) =>
         await this.replacePlayer(gameId, replaceeId, replacementId),
     );
+    this.events.gameReconfigureRequested.subscribe(
+      async ({ gameId }) => await this.reconfigure(gameId),
+    );
   }
 
   async reconfigure(gameId: string) {
@@ -75,6 +78,10 @@ export class GameRuntimeService implements OnModuleInit {
     replaceeId: string,
     replacementId: string,
   ) {
+    if (replaceeId === replacementId) {
+      return;
+    }
+
     const game = await this.gamesService.getById(gameId);
     if (!game.gameServer) {
       return;