+++ src/plugins/discord/services/player-substitution-notifications.service.ts
@@ -1,20 +1,30 @@
 import { Environment } from '@/environment/environment';
 import { Events } from '@/events/events';
+import { SlotStatus } from '@/games/models/slot-status';
 import { GamesService } from '@/games/services/games.service';
-import { Injectable, OnModuleInit } from '@nestjs/common';
-import { Message } from 'discord.js';
+import {
+  CACHE_MANAGER,
+  Inject,
+  Injectable,
+  OnModuleInit,
+} from '@nestjs/common';
+import { Cache } from 'cache-manager';
+import { Snowflake } from 'discord.js';
+import { filter, map } from 'rxjs';
 import { substituteRequest } from '../notifications';
 import { DiscordService } from './discord.service';
 
+const cacheKeyForPlayer = (playerId: string) =>
+  `player-substitute-notifications/${playerId}`;
+
 @Injectable()
 export class PlayerSubstitutionNotificationsService implements OnModuleInit {
-  private notifications = new Map<string, Message>(); // playerId <-> message pairs
-
   constructor(
     private readonly events: Events,
     private readonly discordService: DiscordService,
     private readonly gamesService: GamesService,
     private readonly environment: Environment,
+    @Inject(CACHE_MANAGER) private readonly cache: Cache,
   ) {}
 
   onModuleInit() {
@@ -28,6 +38,27 @@ export class PlayerSubstitutionNotificationsService implements OnModuleInit {
     this.events.playerReplaced.subscribe(
       async ({ replaceeId }) => await this.deleteNotification(replaceeId),
     );
+    this.events.gameChanges
+      .pipe(
+        filter(
+          ({ oldGame, newGame }) =>
+            oldGame.isInProgress() && !newGame.isInProgress(),
+        ),
+        map(({ oldGame }) =>
+          oldGame.slots.filter(
+            (slot) => slot.status === SlotStatus.waitingForSubstitute,
+          ),
+        ),
+      )
+      .subscribe(
+        async (slots) =>
+          await Promise.all(
+            slots.map(
+              async (slot) =>
+                await this.deleteNotification(slot.player.toString()),
+            ),
+          ),
+      );
   }
 
   async notifySubstituteRequested(gameId: string, playerId: string) {
@@ -51,15 +82,24 @@ export class PlayerSubstitutionNotificationsService implements OnModuleInit {
             embeds: [embed],
           })
         : await channel.send({ embeds: [embed] });
-      this.notifications.set(playerId, message);
+      await this.cache.set(cacheKeyForPlayer(playerId), message.id, { ttl: 0 });
     }
   }
 
   async deleteNotification(playerId: string) {
-    const message = this.notifications.get(playerId);
-    if (message) {
-      await message.delete();
-      this.notifications.delete(playerId);
+    const messageId = await this.cache.get(cacheKeyForPlayer(playerId));
+    if (!messageId) {
+      return;
     }
+
+    const message = await this.discordService
+      .getPlayersChannel()
+      .messages.fetch(messageId as Snowflake);
+    if (!message) {
+      return;
+    }
+
+    await message.delete();
+    await this.cache.del(cacheKeyForPlayer(playerId));
   }
 }